requirements.txt
tqdm==4.67.0
omegaconf==2.3.0
hydra-core==1.3.2
rank-bm25==0.2.2
scikit-learn==1.6.1
sacrebleu==2.5.1
evaluate==0.4.3
torch==2.5.1
transformers==4.46.3
openai==1.60.2
